Moved HTTP headers to new file
