+++ packages/jira-adapter/test/deployment/jsp_deployment.test.ts
@@ -19,7 +19,8 @@ import { ElemID, InstanceElement, ObjectType, toChange } from '@salto-io/adapter
 import _ from 'lodash'
 import { mockClient } from '../utils'
 import { deployWithJspEndpoints } from '../../src/deployment/jsp_deployment'
-import JiraClient, { FORCE_ACCEPT_LANGUAGE_HEADERS, JSP_API_HEADERS, PRIVATE_API_HEADERS } from '../../src/client/client'
+import { FORCE_ACCEPT_LANGUAGE_HEADERS, PRIVATE_API_HEADERS, JSP_API_HEADERS } from '../../src/client/headers'
+import JiraClient from '../../src/client/client'
 import { JIRA } from '../../src/constants'
 import { JspUrls } from '../../src/config/config'
